Move generic crypto helpers from 'c.y.s.tls' to 'c.y.s'
