+++ security-utils/src/main/java/com/yahoo/security/SslContextBuilder.java
@@ -1,10 +1,6 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.security;
 
-import com.yahoo.security.tls.KeyManagerUtils;
-import com.yahoo.security.tls.TlsContext;
-import com.yahoo.security.tls.TrustManagerUtils;
-
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
@@ -133,7 +129,7 @@ public class SslContextBuilder {
 
     public SSLContext build() {
         try {
-            SSLContext sslContext = SSLContext.getInstance(TlsContext.SSL_CONTEXT_VERSION);
+            SSLContext sslContext = SSLContext.getInstance("TLS");
             X509ExtendedTrustManager trustManager = this.trustManager != null
                     ? this.trustManager
                     : trustManagerFactory.createTrustManager(trustStoreSupplier.get());
