+++ security-utils/src/main/java/com/yahoo/security/TrustManagerUtils.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.security.tls;
+package com.yahoo.security;
 
 import com.yahoo.security.KeyStoreBuilder;
 import com.yahoo.security.KeyStoreType;
