+++ security-utils/src/main/java/com/yahoo/security/tls/PeerAuthorizerTrustManager.java
@@ -1,6 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.security.tls;
 
+import com.yahoo.security.TrustManagerUtils;
 import com.yahoo.security.X509CertificateUtils;
 
 import javax.net.ssl.SSLEngine;
